+++ NEWS.md
@@ -1,5 +1,9 @@
 # glystats (development version)
 
+## Breaking changes
+
+* All `gly_*_()` matrix and vector interfaces are removed. Pass a `glyexp::experiment()` to the corresponding `gly_*()` function instead. (#12)
+
 # glystats 0.10.1
 
 ## Minor improvements and bug fixes